virtual_interior_designer/requirements.txt
Django==4.2.7
torch==2.0.1
torchvision==0.15.2
opencv-python==4.8.0.76
numpy==1.24.3
segment-anything==1.0
matplotlib==3.7.1
Pillow==10.0.0
scikit-image==0.21.0
omegaconf==2.3.0
pyyaml==6.0.1
easydict==1.10
pytorch-lightning==2.0.6
kornia==0.6.12 
requests==2.31.0 